Fix uncaught exception when sorting columns in Graph Summary
The callback responsible for sorting columns would assume that the Tag
field of the control always held a `DataObjectBase` but this is not
necessarily true.

We now safely check if the type is compatible to avoid an uncaught
exception.

Additionally, I have added code that will make a final attempt to sort
by quality level; though I'm not entirely clear if it actually makes
sense to do so here. If not, please let me know and I can alter/remove
it.

fixes #106