Upgrade Foreman to .NET 9.0, delete old cruft from repo.
This commit upgrades Foreman2 to use .NET 9. This enables us to use
proper dependency management and build with the `dotnet` command line
tools in preparation for having a proper CI build.

Additionally, this enables a newer version of C# and better tooling to
detect code issues.

In order to function correctly under the newer JSON library, the code
has been refactored to avoid deprecated APIs. This could likely benefit
from further refactoring but the scope of this commit is to retain 1:1
behaviour.

I have deleted all files that are irrelevant to the codebase or no
longer needed because they are managed by `dotnet`/`nuget` dependency
management. Since we're using git, they'll be accessible forever anyway.

ForemanTest has also been deleted since it wasn't compilable due to
missing functions. Ideally we'll introduce a test suite in future and
possibly leverage the now-deleted tests, but since it currently serves
no purpose and isn't buildable, remove it.