Apply recommended code cleanup from the C# compiler.
All warnings issued by the compiler have been eliminated. Warnings are
now treated as errors. This also resulted in uncovering a number of bugs
such as a couple of flags encoded in binary notation being duplicates of
eachother and a getter method that did nothing other than call itself
resulting in a stack overflow if invoked.

Most significantly, nullability has been enabled. This has required a
significant amount of work to make the codebase comply with the static
analysis this brings. The way in which JSON is serialised and
deserialised is currently mostly code-driven and quite error prone. I am
planning to refactor this in future but that is out-of-scope for this
commit as its purpose is to do the minimum work necessary to bring the
code base into compliance with the latest C#/.NET standards.

It is naturally possible that during the process of this refactor, bugs
have either been introduced or become surfacable due to less paths being
able to silently fail - I consider this an acceptable trade as it will
contribute to a more robust codebase.