Implement GitHub Actions for building Foreman.
We now use GitHub to build and publish Foreman releases.

This consists of two workflows: snapshot and release.

The snapshot workflow runs for all commits and archives the artifacts so
that you can download them directly from the workflow if you need to
debug anything.

The release workflow only fires when an annotated tag is pushed and it
expects it to be in the format `v1.2.3.4` (replace the numbers as
appropriate). The compiled program will automatically be imbued with the
version number in the tag. Additionally, the action will prepare a draft
release and upload the artifacts into it.